it is not legal to declare bitfield for "char" or "short int" types.

+++ kame/kame/racoon/isakmp.h
@@ -26,7 +26,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  */
-/* YIPS @(#)$Id: isakmp.h,v 1.1 1999/08/08 23:31:22 itojun Exp $ */
+/* YIPS @(#)$Id: isakmp.h,v 1.2 1999/12/22 06:01:16 itojun Exp $ */
 
 /* refer to RFC 2408 */
 
@@ -88,11 +88,11 @@ struct isakmp {
 		u_int8_t ver;
 		struct {
 #if BYTE_ORDER == LITTLE_ENDIAN
-			u_int8_t v_min:4,	/* MnVer */
-				 v_maj:4;	/* MjVer */
+			u_int v_min:4,		/* MnVer */
+				v_maj:4;	/* MjVer */
 #else
-			u_int8_t v_maj:4,	/* MnVer */
-				 v_min:4;	/* MjVer */
+			u_int v_maj:4,		/* MnVer */
+				v_min:4;	/* MjVer */
 #endif
 		} x;
 #define v_number v.ver
